Time: 14 ms (61.59%) | Memory: 56.6 MB (23.18%) - LeetSync

+++ 2581-divide-players-into-teams-of-equal-skill/divide-players-into-teams-of-equal-skill.cpp
@@ -11,7 +11,7 @@ public:
             if(skill[low]+skill[high] != target){
                 return -1;
             }
-            ans+= skill[low]*skill[high];
+            ans+= (long long)skill[low]*skill[high];
             low++;
             high--;
     }
